refactor: moving the frontdoor client
